refactor(components): [steps] switch to script-setup syntax
