+++ packages/components/steps/index.ts
@@ -1,7 +1,7 @@
 import { withInstall, withNoopInstall } from '@element-plus/utils'
 
-import Steps from './src/index.vue'
-import Step from './src/item.vue'
+import Steps from './src/steps.vue'
+import Step from './src/step.vue'
 
 export const ElSteps = withInstall(Steps, {
   Step,
